!516 fix(process): harden exit cleanup and wait visibility
* fix: harden mm user final release contract
* fix: type process-domain exit-state reads
* fix: make signal send interruption decision atomic
* fix: update TEE TA context for all TEE execs
* Fix rebase integration fallout
* Avoid mm invalidate handle relock
* Refine process mm lifetime capabilities
* Refine process exit publication model
* kprocess: reserve publication slots atomically
* kprocess: name wait parent contract explicitly
* kprocess: split process state modules
* kprocess: clarify publication slot reservation
* kprocess: type controlling terminal slot
* kprocess: require domain token for publication helpers
* kprocess: require domain token for tree mutation
* ksignal: document dequeue observer context
* kprocess: hide process timer manager lock
* kprocess: clarify published process lookups
* sched: avoid stale policy in setparam
* kprocess: serialize publication slot commits
* kprocess: roll back failed fork runtime attach
* kprocess: type publication rollback effects
* kprocess: bind thread publication slots
* kprocess: tighten process runtime state model
* kprocess: send child-exit siginfo payload
* ksignal: choose child-exit target at commit
* kprocess: prepare child-exit signal outside tree lock
* kprocess: tighten publication slot invariants
* kprocess: serialize process publication and reap
* kspin: add irq-safe spin rwlock
* fix ai review
* fix(process): harden exit cleanup and wait visibility

+++ core/ksyscall/src/task/cpu_time.rs
@@ -18,8 +18,8 @@ pub fn sys_times(tms: UserPtr<Tms>) -> KResult<isize> {
     tms.write_vm(Tms {
         tms_utime: utime,
         tms_stime: stime,
-        tms_cutime: ns2t(child_utime_ns as u64) as usize,
-        tms_cstime: ns2t(child_stime_ns as u64) as usize,
+        tms_cutime: ns2t(child_utime_ns) as usize,
+        tms_cstime: ns2t(child_stime_ns) as usize,
     })?;
     Ok(ns2t(monotonic_time_nanos()) as _)
 }
